+++ src/layouts/View/components/Navbar/utils.ts
@@ -1,3 +1,10 @@
+import { env } from "@config";
+import {
+  readDailyModeOutcomeForDate,
+  resolvePlayableWordleModeId,
+  WORDLE_MODE_IDS,
+  type WordleModeId,
+} from "@domain/wordle";
 import { TOP_TEN_LIMIT } from "./constants";
 
 export const getScoreboardToneClassName = (rank: number | null): string => {
@@ -11,3 +18,27 @@ export const getScoreboardToneClassName = (rank: number | null): string => {
 
   return "border border-neutral-300 bg-neutral-200/80 text-neutral-600 hover:bg-neutral-300/90";
 };
+
+export const normalizePathname = (pathname: string): string => {
+  if (pathname.length <= 1) {
+    return pathname;
+  }
+
+  return pathname.endsWith("/") ? pathname.slice(0, -1) : pathname;
+};
+
+export const resolveNavbarPlayableModeId = (
+  modeId: WordleModeId,
+): WordleModeId => {
+  if (modeId === WORDLE_MODE_IDS.LIGHTNING && !env.lightningModeEnabled) {
+    return WORDLE_MODE_IDS.CLASSIC;
+  }
+
+  return resolvePlayableWordleModeId(modeId);
+};
+
+export const hasResolvedDailyOutcomeForToday = (
+  playerCode?: string | null,
+): boolean =>
+  readDailyModeOutcomeForDate(playerCode) !== null ||
+  readDailyModeOutcomeForDate() !== null;
