+++ src/views/Play/components/Dialogs/GameplayTourDialog/constants.ts
@@ -1,0 +1,6 @@
+export const HIGHLIGHT_BOX_SHADOW = "0 0 0 9999px rgba(0, 0, 0, 0.62)";
+export const HIGHLIGHT_BORDER_COLOR = "rgba(59, 130, 246, 0.9)";
+export const HIGHLIGHT_PADDING_PX = 10;
+export const HIGHLIGHT_VIEWPORT_MARGIN_PX = 8;
+export const TOUR_DIALOG_VIEWPORT_MARGIN_PX = 16;
+export const GAMEPLAY_TOUR_DIALOG_TITLE_ID = "play-gameplay-tour-dialog-title";
