+++ src/views/Play/types.ts
@@ -1,0 +1,5 @@
+import type { WordleModeId } from "@domain/wordle";
+
+export type PlayProps = {
+  modeId?: WordleModeId;
+};
